test_metatesting.py
import os
import shutil
import unittest
from unittest.mock import MagicMock, patch
from logger import Logger
from testing import Testing, TestSummary
from metatesting import MetaTesting
from functionmapping import FunctionMapping

class TestMetaTesting(unittest.TestCase):
    def setUp(self):
        self.test_directory = "test_artifacts"
        os.makedirs(self.test_directory, exist_ok=True)
        self.logger = Logger()
        self.testing = Testing(self.logger, FunctionMapping())
        self.metatesting = MetaTesting(self.logger, self.testing)

    def tearDown(self):
        shutil.rmtree(self.test_directory, ignore_errors=True)

    def test_generateTestFunctionDefinitions(self):
        function_definitions = self.metatesting.generateTestFunctionDefinitions()
        self.assertIsInstance(function_definitions, list)
        self.assertTrue(all(isinstance(item, list) for item in function_definitions))
        # Add more specific assertions based on the expected function definitions

    def test_createTestModule(self):
        module_path = os.path.join(self.test_directory, "test_module.py")
        function_definitions = ["def test_function():", "    pass"]
        self.metatesting.createTestModule(module_path, function_definitions)
        self.assertTrue(os.path.exists(module_path))
        with open(module_path, "r") as file:
            content = file.read()
            self.assertEqual(content, "\n".join(function_definitions))

    def test_generateTestInputs(self):
        test_inputs = self.metatesting.generateTestInputs("test_function_1")
        self.assertIsInstance(test_inputs, list)
        self.assertTrue(all(isinstance(item, tuple) for item in test_inputs))
        # Add more specific assertions based on the expected test inputs

    def test_runTestFunctions(self):
        # Create sample test modules
        test_module_paths = []
        for i in range(2):
            module_path = os.path.join(self.test_directory, f"test_module_{i}.py")
            function_definitions = [
                f"def test_function_{i}():",
                f"    return {i}"
            ]
            self.metatesting.createTestModule(module_path, function_definitions)
            test_module_paths.append(module_path)

        log_file_paths = self.metatesting.runTestFunctions(test_module_paths)
        self.assertIsInstance(log_file_paths, list)
        self.assertEqual(len(log_file_paths), 2)
        # Add more specific assertions based on the expected log file paths and contents

    def test_cleanUpTestArtifacts(self):
        file_path = os.path.join(self.test_directory, "test_file.txt")
        directory_path = os.path.join(self.test_directory, "test_dir")
        os.makedirs(directory_path)
        with open(file_path, "w") as file:
            file.write("Test content")

        self.metatesting.cleanUpTestArtifacts(file_path)
        self.assertFalse(os.path.exists(file_path))

        self.metatesting.cleanUpTestArtifacts(directory_path)
        self.assertFalse(os.path.exists(directory_path))

    def test_generateTestModules(self):
        test_module_paths = self.metatesting.generateTestModules(self.test_directory)
        self.assertIsInstance(test_module_paths, list)
        self.assertTrue(all(os.path.exists(path) for path in test_module_paths))
        # Add more specific assertions based on the expected test module paths and contents

    def test_runSelfTests(self):
        test_summary = self.metatesting.runSelfTests(self.test_directory)
        self.assertIsInstance(test_summary, TestSummary)
        # Add more specific assertions based on the expected test summary values
        self.assertTrue(os.path.exists(self.test_directory))
        self.logger.logOutput.assert_called()
        self.testing.runTestSuite.assert_called()

    @patch("metatesting.MetaTesting")
    def test_integration(self, mock_metatesting):
        mock_metatesting.return_value.runSelfTests.return_value = TestSummary(passed=2, failed=0, skipped=0)
        metatesting = MetaTesting(self.logger, self.testing)
        test_summary = metatesting.runSelfTests(self.test_directory)
        self.assertIsInstance(test_summary, TestSummary)
        self.assertEqual(test_summary.passed, 2)
        self.assertEqual(test_summary.failed, 0)
        self.assertEqual(test_summary.skipped, 0)
        # Add more specific assertions based on the expected test summary values and side effects
        self.logger.logOutput.assert_called()
        self.testing.runTestSuite.assert_called()

    def test_error_handling(self):
        # Test invalid test directory path
        with self.assertRaises(FileNotFoundError):
            self.metatesting.runSelfTests("invalid/path")

        # Test invalid module path
        with self.assertRaises(FileNotFoundError):
            self.metatesting.runTestFunctions(["invalid/module/path"])

        # Test missing dependencies
        with self.assertRaises(TypeError):
            MetaTesting(None, None)

        # Add more test cases for error handling and edge cases

if __name__ == "__main__":
    unittest.main()
